fix(dropdown): renamed default-value to value and refactored set/removeValue logic
Renamed default-value to value to allow us to reflect the value onto the host element when a
selection is made. Also refactored the logic in the set/removeValue functions to "recheck" which
it's correct child elements are. This is to allow for dynamic rendering of children.

BREAKING CHANGE: default-value was renamed to value